Use float for length

+++ roundtrip/app.py
@@ -12,7 +12,7 @@ def index():
         latitude = request.form["latitude"]
         length = request.form["length"]
 
-        gpx_data = generate_gpx([float(longitude), float(latitude)], int(length) * 1000)
+        gpx_data = generate_gpx([float(longitude), float(latitude)], float(length) * 1000)
         return Response(gpx_data, mimetype="application/gpx+xml")
 
     return render_template("index.html")
